consolidate clickButton func
